test(generation): make the kumiko split three scenario modules
The first attempt kept one scenarios/kumiko.ts and gave it three test file
pairs, which binGenerator.scenarioCoverage.test.ts correctly rejected: it
enforces one module to exactly one scenario file and one export file, and
asserts the export file count equals the module count. That guard exists so a
module wired under the wrong name cannot silently double-run or drop
scenarios.

Splits the module itself into kumikoPatterns, kumikoWrapping and
kumikoComposition, so each of the six test files has its own module and the
convention holds. ALL_SIDES_OFF is exported from kumikoWrapping and shared,
mirroring how wallPatterns already exports assertRemovesMaterial.

index.ts spreads the three, so ALL_SCENARIOS is unchanged. 30 tests still
run: 152s wall against 486s of test time, parallelism two files could not
express.
