+++ src/features/generation/worker/generators/binGenerator.export.kumikoComposition.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
 import { runExportIntegrity } from './__kernel-tests__/exportIntegrityRunner';
-import { kumikoComposition } from './scenarios/kumiko';
+import { kumikoComposition } from './scenarios/kumikoComposition';
 
 runExportIntegrity(kumikoComposition);
