+++ src/features/generation/worker/generators/binGenerator.export.kumikoPatterns.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
 import { runExportIntegrity } from './__kernel-tests__/exportIntegrityRunner';
-import { kumikoPatterns } from './scenarios/kumiko';
+import { kumikoPatterns } from './scenarios/kumikoPatterns';
 
 runExportIntegrity(kumikoPatterns);
