+++ src/features/generation/worker/generators/binGenerator.export.kumikoWrapping.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
 import { runExportIntegrity } from './__kernel-tests__/exportIntegrityRunner';
-import { kumikoWrapping } from './scenarios/kumiko';
+import { kumikoWrapping } from './scenarios/kumikoWrapping';
 
 runExportIntegrity(kumikoWrapping);
