+++ src/features/generation/worker/generators/binGenerator.scenario.kumikoComposition.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
 import { runScenarios } from './__kernel-tests__/scenarioRunner';
-import { kumikoComposition } from './scenarios/kumiko';
+import { kumikoComposition } from './scenarios/kumikoComposition';
 
 runScenarios(kumikoComposition);
