+++ src/features/generation/worker/generators/binGenerator.scenario.kumikoPatterns.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
 import { runScenarios } from './__kernel-tests__/scenarioRunner';
-import { kumikoPatterns } from './scenarios/kumiko';
+import { kumikoPatterns } from './scenarios/kumikoPatterns';
 
 runScenarios(kumikoPatterns);
