+++ src/features/generation/worker/generators/binGenerator.scenario.kumikoWrapping.test.ts
@@ -1,5 +1,5 @@
 // @vitest-environment node
 import { runScenarios } from './__kernel-tests__/scenarioRunner';
-import { kumikoWrapping } from './scenarios/kumiko';
+import { kumikoWrapping } from './scenarios/kumikoWrapping';
 
 runScenarios(kumikoWrapping);
