+++ src/features/generation/worker/generators/scenarios/kumikoWrapping.ts
@@ -17,7 +17,7 @@ import { DEFAULT_PATTERN_SCALE } from '@/shared/types/bin';
 import { defineScenario } from '../__kernel-tests__/scenarioTypes';
 import type { ScenarioCase } from '../__kernel-tests__/scenarioTypes';
 import type { MeshData } from '@/features/generation/bridge/types';
-import type { BinParams, WallPatternType } from '@/shared/types/bin';
+import type { BinParams } from '@/shared/types/bin';
 import { deriveDimensions } from '../pipeline/context';
 import { TOP_KEEP_OUT, BOTTOM_SOLID_SKIRT } from '../wallPatterns';
 import { SOCKET_HEIGHT, BOX_CORNER_RADIUS } from '../generatorConstants';
@@ -26,7 +26,7 @@ import { generateKumikoLattice, KUMIKO_BASE_CELL_SIZE } from '../patterns/kumiko
 import { MITSUKUDE_DEF } from '../patterns/kumiko/mitsukude';
 import { assertRemovesMaterial } from './wallPatterns';
 
-const ALL_SIDES_OFF = {
+export const ALL_SIDES_OFF = {
   ...DEFAULT_BIN_PARAMS.walls,
   enabled: false,
   front: DISABLED_WALL_CUTOUT,
@@ -247,40 +247,6 @@ function assertCornerDiagonalsPresent(mesh: MeshData, params: BinParams): void {
   expect(missing, missing.join('\n')).toEqual([]);
 }
 
-/** Compact per-pattern case: valid geometry + material removed vs solid twin. */
-function kumikoPatternCase(pattern: WallPatternType): ScenarioCase {
-  return defineScenario('kumiko', `${pattern} carves a 1×1×6 bin`, {
-    assert: 'structural',
-    timeout: 180_000,
-    params: {
-      width: 1,
-      depth: 1,
-      height: 6,
-      wallPattern: { enabled: true, pattern, scale: 0.5 },
-      walls: ALL_SIDES_OFF,
-    },
-    compareWith: {
-      params: {
-        width: 1,
-        depth: 1,
-        height: 6,
-        wallPattern: { enabled: false, pattern, scale: 0.5 },
-        walls: ALL_SIDES_OFF,
-      },
-      assert: assertRemovesMaterial,
-    },
-  });
-}
-
-// One case per kumiko pattern: each must remove material from a 1×1×6 bin.
-export const kumikoPatterns: ScenarioCase[] = [
-  kumikoPatternCase('goma'),
-  kumikoPatternCase('sakura'),
-  kumikoPatternCase('rindo'),
-  kumikoPatternCase('mikado'),
-  kumikoPatternCase('tsumiishi-kikko'),
-];
-
 // How a pattern reaches corners, and which walls it covers.
 export const kumikoWrapping: ScenarioCase[] = [
   defineScenario('kumiko', 'mitsukude keeps both diagonal families on corners (1×1×4, bold)', {
@@ -386,80 +352,3 @@ export const kumikoWrapping: ScenarioCase[] = [
     },
   }),
 ];
-
-// Kumiko composed with other bin features.
-export const kumikoComposition: ScenarioCase[] = [
-  defineScenario('kumiko', 'mitsukude composes with a front wall cutout', {
-    assert: 'structural',
-    timeout: 180_000,
-    params: {
-      width: 2,
-      depth: 1,
-      height: 6,
-      wallPattern: { enabled: true, pattern: 'mitsukude', scale: 0.5 },
-      walls: {
-        ...ALL_SIDES_OFF,
-        enabled: true,
-        front: { ...DISABLED_WALL_CUTOUT, enabled: true, width: 60, depth: 50 },
-      },
-    },
-  }),
-  defineScenario('kumiko', 'mitsukude composes with handles', {
-    assert: 'structural',
-    timeout: 180_000,
-    params: {
-      width: 2,
-      depth: 1,
-      height: 6,
-      wallPattern: { enabled: true, pattern: 'mitsukude', scale: 0.5 },
-      walls: ALL_SIDES_OFF,
-      handles: {
-        ...DEFAULT_BIN_PARAMS.handles,
-        enabled: true,
-        front: { ...DEFAULT_BIN_PARAMS.handles.front, enabled: true },
-      },
-    },
-  }),
-  defineScenario('kumiko', 'mitsukude composes with 2×2 compartment dividers', {
-    assert: 'structural',
-    timeout: 180_000,
-    params: {
-      width: 2,
-      depth: 2,
-      height: 6,
-      wallPattern: { enabled: true, pattern: 'mitsukude', scale: 0.5 },
-      walls: ALL_SIDES_OFF,
-      compartments: { cols: 2, rows: 2, cells: [0, 1, 2, 3], thickness: 0.8 },
-    },
-  }),
-  defineScenario('kumiko', 'mitsukude on a half-grid 1.5×1×6 bin with magnets', {
-    assert: 'structural',
-    timeout: 180_000,
-    params: {
-      width: 1.5,
-      depth: 1,
-      height: 6,
-      wallPattern: { enabled: true, pattern: 'mitsukude', scale: 0.5 },
-      walls: ALL_SIDES_OFF,
-      base: { ...DEFAULT_BIN_PARAMS.base, style: 'magnet' },
-    },
-  }),
-  defineScenario('kumiko', 'mitsukude with asymmetric overhang', {
-    assert: 'structural',
-    timeout: 180_000,
-    params: {
-      width: 2,
-      depth: 2,
-      height: 6,
-      wallPattern: { enabled: true, pattern: 'mitsukude', scale: 0.5 },
-      walls: ALL_SIDES_OFF,
-      overhang: { ...DEFAULT_BIN_PARAMS.overhang, left: 4, right: 0, front: 2, back: 0 },
-    },
-  }),
-];
-
-// Aggregate for scenarios/index.ts. The three groups above each get their own
-// test file: at ~13s per case these fifteen were a single 200s file, and Vitest
-// never parallelizes within a file, so it floored the whole generators shard no
-// matter how the suite was sharded. Same fix as #2882.
-export const kumiko: ScenarioCase[] = [...kumikoPatterns, ...kumikoWrapping, ...kumikoComposition];
